Modified and implemented sign out and cart features

+++ src/app/sign-out/sign-out.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { Router } from '@angular/router';
 import { CookieService } from 'ngx-cookie-service';
 
 @Component({
@@ -8,17 +9,31 @@ import { CookieService } from 'ngx-cookie-service';
 })
 export class SignOutComponent implements OnInit {
 
-  constructor(private cookieService: CookieService) { }
+  username: String = "";
+  buttoninfo: String = "Sign-in";
+
+  constructor(private cookieService: CookieService, private router: Router) { }
 
   onSubmit() {
 
-    this.cookieService.deleteAll();
-    window.location.reload();
+    if (this.username === "") {
+      this.router.navigate(['/login']);
+    } else {
+      this.cookieService.deleteAll();
+      window.location.reload();
+    }
+
 
 
   }
 
   ngOnInit(): void {
+    this.username = this.cookieService.get("username");
+    if (this.username === "") {
+      this.buttoninfo = "Sign-in";
+    } else {
+      this.buttoninfo = "Sign-out";
+    }
   }
 
 }
